Debounce parse_wdl() per URI in language server
Previously, parse_wdl() function in the language server
was debounced globally for all requests.

This meant that it could not parse >1 file at once.

This commit adds an extra argument for @debounce() decorator
to define which argument of the debounced function
it should consider as the "debouncing ID".

It then debounces calls to that function per each ID
(by maintaining a dict of timers keyed on the ID).

In our use case, parse_wdl() will thus be debounced
on the "uri" argument, i.e. per each file.